[Feat] Nest sql DDL 추가
