+++ keewe-domain/src/main/java/ccc/keewedomain/domain/nest/Comment.java
@@ -1,5 +1,6 @@
 package ccc.keewedomain.domain.nest;
 
+import ccc.keewedomain.domain.common.BaseTimeEntity;
 import ccc.keewedomain.domain.user.Profile;
 
 import javax.persistence.*;
@@ -10,18 +11,12 @@ import java.util.List;
 import static javax.persistence.FetchType.LAZY;
 
 @Entity
-public class Comment {
+@Table(name = "comment")
+public class Comment extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "comment_id")
     private Long id;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "parent_comment_id", nullable = true)
-    private Comment parent;
-
-    @OneToMany(mappedBy = "parent", fetch = LAZY)
-    private List<Comment> replies = new ArrayList<>();
 
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "post_id", nullable = false)
@@ -31,6 +26,16 @@ public class Comment {
     @JoinColumn(name = "writer_id", nullable = false)
     private Profile writer;
 
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "parent_comment_id", nullable = true)
+    private Comment parent;
+
+    @OneToMany(mappedBy = "parent", fetch = LAZY)
+    private List<Comment> replies = new ArrayList<>();
+
     @Column(name = "content", nullable = false)
     private String content;
+
+    @Column(name = "deleted", nullable = false)
+    private boolean deleted = false;
 }
