+++ keewe-domain/src/main/java/ccc/keewedomain/domain/nest/CommentLike.java
@@ -1,5 +1,7 @@
 package ccc.keewedomain.domain.nest;
 
+import ccc.keewedomain.domain.common.BaseTimeEntity;
+import ccc.keewedomain.domain.nest.id.CommentLikeId;
 import ccc.keewedomain.domain.user.Profile;
 
 import javax.persistence.*;
@@ -7,8 +9,9 @@ import javax.persistence.*;
 import static javax.persistence.FetchType.LAZY;
 
 @Entity
+@Table(name = "comment_like")
 @IdClass(CommentLikeId.class)
-public class CommentLike {
+public class CommentLike extends BaseTimeEntity {
 
     @Id
     @ManyToOne(fetch = LAZY)
