+++ keewe-domain/src/main/java/ccc/keewedomain/domain/nest/id/ChoiceId.java
@@ -1,0 +1,13 @@
+package ccc.keewedomain.domain.nest.id;
+
+import ccc.keewedomain.domain.nest.Candidate;
+import ccc.keewedomain.domain.user.Profile;
+import lombok.EqualsAndHashCode;
+
+import java.io.Serializable;
+
+@EqualsAndHashCode
+public class ChoiceId implements Serializable {
+    private Candidate candidate;
+    private Profile profile;
+}
