+++ keewe-domain/src/main/java/ccc/keewedomain/domain/nest/id/PostLikeId.java
@@ -1,12 +1,14 @@
-package ccc.keewedomain.domain.nest;
+package ccc.keewedomain.domain.nest.id;
 
+import ccc.keewedomain.domain.nest.Post;
 import ccc.keewedomain.domain.user.Profile;
 import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
 
+
 @EqualsAndHashCode
-public class ChoiceId implements Serializable {
-    private Candidate candidate;
+public class PostLikeId implements Serializable {
     private Profile profile;
+    private Post post;
 }
